Fix ribbon-tail hiding on mobile viewports. :nut_and_bolt:
The ribbon-tail mixin needed to also resize (and so re-offset) the tail
containers when in mobile viewports. Not being rendered, the tail
containers also need to clip their content.

Refactor the mixin along the way to flip that change by defaulting to
the mobile state and adding non-mobile styles on top, including animation.